[BUGFIX] Flaws in copying a page with localized content elements
Copying a page with localized content elements in a workspace
has some flaws. In a workspace a duplicated record will end up
in a new-placeholder and the specific version record.

The tranlation pointer field of the new-placeholder still
contains the unmodified value of the original record, the
specific version record is fine however.

The reason for that is the special handling for cloning a page
in DataHandler. Each record of a particular table is copied in
reverse order. This way, the localization is created earlier
than the record of the default language.

This change is about skipping the direct copyRecord() call for
records that will be processed inside another copyRecord() call
for the record of the default language.

Resolves: #61179
Releases: master, 6.2
Change-Id: I6647b02095343756d3a41b0c0e985b4777429cd8