[BUGFIX] Child sorting is lost after publishing workspace data
If sorting of IRRE child records is changed in a workspace and
published, the sorting of child records is not published along.

This behaviour could be solved by using proper move placeholders
for IRRE child records as well, but conflicts with the
possibility to define a custom sorting field in the parent
entity of the relation construct. Besides that, sorting
values are not published in any case. That's why a
post-processing of IRRE children is required for sorting.

Resolves: #61750
Releases: master, 6.2
Change-Id: I80405299103b039b192626e20fcd74a05f96795f