[TASK] Add IE 11 detection to ExtJS
ExtJS 3.4.1 used in the TYPO3 Core does not ship browser detection for
Internet Explorer 11. Because IE 11 uses "Gecko" in the user-agent
string to be detected as a modern browser, ExtJS treats IE11 as Gecko
browser which leads to various problems with JavaScript features that
are not part of IE.

For financial and licensing reasons, TYPO3 cannot update to ExtJS 3.4.2
which includes the detection. Therefore, this change adds IE 11 detection
to ExtJS. ExtJS now returns true for "Ext.isIE" and "Ext.isIE11" and
false for "Ext.isGecko".

Resolves: #55774
Resolves: #55775
Resolves: #56192
Releases: master,6.2
Change-Id: Id86cde375aeba3a314db6baadad4446f38b23eb0
Reviewed-on: http://review.typo3.org/37766
Reviewed-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Tested-by: Alexander Stehlik <alexander.stehlik@gmail.com>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>