[WIP][FEATURE] Add workspace element filter settings
The workspace configuration is extended by an element filter
setting that allows to define record visibility in the workspace
module per stage. This way, it can be defined, that either
workspace owner, members or editors/creators of a particular
record are allowed to see elements in the module - depending
on the setting for the element's workspace stage.

Resolves: #65720
Releases: master
Change-Id: I4876ba626ec8c2beef20d50b962939540a332a33