[SECURITY] Regenerate session id upon login if needed
When authenticating as a frontend user with a previously
present anonymous session, the session id is not regenerated
which leads to a possible session fixation.

This is now fixed by re-generating a new id
when a user is just authenticated but no
new session id is generated during this process.

Resolves: #59258
Change-Id: I860fed8450f14122ace9e2ff203d9421e5df2bf8