[SECURITY] Add hook to implement login protection methods
Currently only the backend login is protected with an implement sleep
time after login failure. This patch adds a new hook which can be used
to implement a protection functionality (e.g. for detecting brute force)
and moves the sleep time as default protection in the abstract user
authentication class.

Resolves: #59231
Change-Id: I2ca7383d356cea920b00614ee8bcd38adda8fa9e